Extract non-DEM source file bake emitter
